# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: profile
    description: Authenticated profile and portfolio operations
  - name: uploads
    description: Authenticated image uploads
  - name: payments
    description: Server-side Paystack operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /profile/me:
    get:
      operationId: getMyProfile
      tags: [profile]
      summary: Get the current user's profile
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Profile and portfolio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      operationId: updateMyProfile
      tags: [profile]
      summary: Create or update the current user's profile
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Saved profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /uploads/profile:
    post:
      operationId: uploadProfileImage
      tags: [uploads]
      summary: Upload a profile or portfolio image
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Public image URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
  /payments/paystack/verify/{reference}:
    get:
      operationId: verifyPaystackPayment
      tags: [payments]
      summary: Verify a Paystack transaction
      security: [{ bearerAuth: [] }]
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Paystack verification response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: Authentication required
    BadRequest:
      description: Invalid request
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ProfileInput:
      type: object
      required: [displayName, role]
      properties:
        displayName: { type: string }
        phone: { type: string }
        city: { type: string }
        bio: { type: string }
        role:
          type: string
          enum: [customer, artisan, professional, business, admin]
        skills:
          type: array
          items: { type: string }
        yearsExperience: { type: number, minimum: 0 }
        hourlyRateNgn: { type: number, minimum: 0 }
        availability: { type: string }
        avatarUrl: { type: string, format: uri }
        portfolio:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioInput"
    PortfolioInput:
      type: object
      required: [title]
      properties:
        title: { type: string }
        description: { type: string }
        imageUrl: { type: string, format: uri }
        projectUrl: { type: string, format: uri }
        sortOrder: { type: integer }
    ProfileResponse:
      type: object
      properties:
        profile:
          type: object
          additionalProperties: true
        portfolio:
          type: array
          items:
            type: object
            additionalProperties: true
    UploadResponse:
      type: object
      required: [path, url]
      properties:
        path: { type: string }
        url: { type: string, format: uri }