# lib/api-zod/src/generated/types/profileResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileResponsePortfolioItem } from './profileResponsePortfolioItem';
import type { ProfileResponseProfile } from './profileResponseProfile';

export interface ProfileResponse {
  profile?: ProfileResponseProfile;
  portfolio?: ProfileResponsePortfolioItem[];
}